refactor(server): extract worker auth helpers
