+++ server/proliferate/server/cloud/backfill/api.py
@@ -12,7 +12,7 @@ from proliferate.server.cloud.backfill.models import (
 )
 from proliferate.server.cloud.backfill.service import record_worker_backfill
 from proliferate.server.cloud.errors import CloudApiError, raise_cloud_error
-from proliferate.server.cloud.worker.service import authenticate_worker
+from proliferate.server.cloud.worker.auth import authenticate_worker
 
 router = APIRouter(prefix="/worker/backfill", tags=["cloud-worker-backfill"])
